fix: cart style
